Split updater wifi modules
