+++ apps/server/tests/use_cases/updates/test_update_manager_models.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-from pathlib import Path
-
 from vibesensor.use_cases.updates.models import (
     UpdateIssue,
     UpdateJobStatus,
@@ -13,7 +11,6 @@ from vibesensor.use_cases.updates.venv_paths import (
     is_reinstall_venv_ready,
     reinstall_python_executable,
 )
-from vibesensor.use_cases.updates.wifi import parse_wifi_diagnostics
 
 
 class TestUpdateJobStatus:
@@ -91,61 +88,3 @@ class TestSanitizeLogLine:
     def test_normal_line_unchanged(self) -> None:
         line = "Hotspot restored on attempt 1"
         assert sanitize_log_line(line) == line
-
-
-class TestParseWifiDiagnostics:
-    def test_no_log_dir(self, tmp_path) -> None:
-        assert parse_wifi_diagnostics(str(tmp_path / "nonexistent")) == []
-
-    def test_summary_failure(self, tmp_path) -> None:
-        log_dir = tmp_path / "wifi"
-        log_dir.mkdir()
-        (log_dir / "summary.txt").write_text("status=FAILED\nrc=22\n")
-        issues = parse_wifi_diagnostics(str(log_dir))
-        assert any("failure" in issue.message.lower() for issue in issues)
-
-    def test_summary_timeout_case_insensitive(self, tmp_path) -> None:
-        log_dir = tmp_path / "wifi"
-        log_dir.mkdir()
-        (log_dir / "summary.txt").write_text("status=timeout\n")
-        issues = parse_wifi_diagnostics(str(log_dir))
-        assert any("timeout" in (issue.detail or "").lower() for issue in issues)
-
-    def test_summary_password_not_leaked(self, tmp_path) -> None:
-        log_dir = tmp_path / "wifi"
-        log_dir.mkdir()
-        (log_dir / "summary.txt").write_text("status=FAILED psk=hunter2\n")
-        issues = parse_wifi_diagnostics(str(log_dir))
-        for issue in issues:
-            assert "hunter2" not in (issue.detail or "")
-
-    def test_hotspot_log_errors(self, tmp_path) -> None:
-        log_dir = tmp_path / "wifi"
-        log_dir.mkdir()
-        (log_dir / "hotspot.log").write_text(
-            "2024-01-01 INFO normaline\n2024-01-01 ERROR something failed\n2024-01-01 INFO ok\n",
-        )
-        issues = parse_wifi_diagnostics(str(log_dir))
-        assert any(
-            "error" in issue.detail.lower() or "failed" in issue.detail.lower() for issue in issues
-        )
-
-    def test_password_not_leaked_in_diagnostics(self, tmp_path) -> None:
-        log_dir = tmp_path / "wifi"
-        log_dir.mkdir()
-        (log_dir / "hotspot.log").write_text("ERROR psk=hunter2 failed\n")
-        issues = parse_wifi_diagnostics(str(log_dir))
-        for issue in issues:
-            assert "hunter2" not in issue.detail
-            assert "hunter2" not in issue.message
-
-    def test_read_errors_are_ignored(self, tmp_path, monkeypatch) -> None:
-        log_dir = tmp_path / "wifi"
-        log_dir.mkdir()
-        (log_dir / "summary.txt").write_text("status=FAILED\n")
-
-        def raise_oserror(*_args, **_kwargs) -> str:
-            raise OSError("boom")
-
-        monkeypatch.setattr(Path, "read_text", raise_oserror)
-        assert parse_wifi_diagnostics(str(log_dir)) == []
